major refactoring again
