+++ pySW4/core/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Core routines of pySW4
+"""
+from .config import read_input_file
+from .image import Image, read_image
