+++ pySW4/core/header.py
