+++ pySW4/plotting/__init__.py
@@ -1,37 +1,7 @@
 # -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-#  Purpose: Convenience imports for SeisPy
-#   Author: Shahar Shani-Kadmiel
-#           kadmiel@post.bgu.ac.il
-#
-# Copyright ©(C) 2012-2014 Shahar Shani-Kadmiel
-# This code is distributed under the terms of the GNU General Public License
-# -----------------------------------------------------------------------------
 """
-
-   _____      _      ____
-  / ___/___  (_)____/ __ \__  __
-  \__ \/ _ \/ / ___/ /_/ / / / /
- ___/ /  __/ (__  ) ____/ /_/ /
-/____/\___/_/____/_/    \__, /
-                       /____/
-
-
-SeisPy: A Python Toolbox for processing Seismic-wave propagation simulations
-============================================================================
-
-SeisPy is an open-source project dedicated to provide a Python framework for
-processing numerical simulations of seismic-wave propagation in all phases of
-the task (preprocessing, post-processing and runtime visualization).
-
-:copyright:
-    Shahar Shani-Kadmiel (kadmiel@post.bgu.ac.il)
-    and
-    Omry Volk (enter email here)
-:license:
-    GNU GENERAL PUBLIC LICENSE Version 3
-    (http://www.gnu.org/copyleft/gpl.html)
-    See ./LICENSE.txt
+Plotting routines of pySW4
 """
-from drape import *
-from utils import *
+from .drape import drape_plot
+from .image import create_image_plots
+from .waveforms import create_seismogram_plots
